% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtraffic.R
\docType{data}
\name{airtraffic}
\alias{airtraffic}
\title{Monthly Airline Traffic in the United States}
\format{
An object of class \code{data.frame} with 249 rows and 3 columns.
}
\source{
https://www.kaggle.com/datasets/yyxian/u-s-airline-traffic-data
}
\usage{
airtraffic
}
\description{
Monthly international airline passengers data between 2003 and 2023.
}
\examples{
data(airtraffic)
head(airtraffic)
}
\references{
Yan, Y. (2023). U.S. Airline Traffic Data (2003–2023). Retrieved from https://www.kaggle.com/datasets/yyxian/u-s-airline-traffic-data (Accessed: June 16, 2024).
}
\keyword{datasets}
