% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clusterGraph}
\alias{clusterGraph}
\title{Remove inter-cluster edges from a graph}
\usage{
clusterGraph(graph, membership)
}
\arguments{
\item{graph}{An object of class 'graph' from the igraph package.}

\item{membership}{A vector of integers of length N (see \code{\link{constructClusters}})}
}
\value{
An object of class 'graph' from the igraph package, the input graph with inter-cluster edges removed
}
\description{
Remove inter-cluster edges from a graph
}
\examples{
set.seed(1)
coords = data.frame(lon = rnorm(50), lat = rnorm(50))
g = constructGraph(coords, 4)
clust_membership = constructClusters(g, 5, minclust = 3)$membership
clust_g = clusterGraph(g, clust_membership)
plot(clust_g, layout = as.matrix(coords), vertex.color = clust_membership, edge.arrow.mode = 0)
}
