% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{constructClusters}
\alias{constructClusters}
\title{Construct a cluster membership list for a graph}
\usage{
constructClusters(graph, nclust, minclust = NULL)
}
\arguments{
\item{graph}{An object of class 'graph' from the igraph package. The graph must have weights for each edge}

\item{nclust}{An integer, the number of different clusters to assign points to. Must be at most N (the number of vertices in graph)}

\item{minclust}{An integer, the smallest allowable cluster size. By default, one-tenth the ratio of vertices to nclust.}
}
\value{
A list containing two elements:
'clustered_graph': The input graph with inter-cluster edges removed
'spanning_forest': The input graph with inter-cluster edges removed, and every cluster induced subgraph is a spanning tree
'membership': A vector of integers of length N with nclust unique integers which map each vertex to a cluster
}
\description{
Construct a cluster membership list for a graph
}
\examples{
set.seed(1)
coords = data.frame(lon = rnorm(50), lat = rnorm(50))
g = constructGraph(coords, 4)
clust_out = constructClusters(g, 5, minclust = 3)
plot(clust_out$spanning_forest, layout = as.matrix(coords), vertex.color = clust_out$membership, edge.arrow.mode = 0)
}
