% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{constructGraph}
\alias{constructGraph}
\title{Construct a Graph Object from Coordinate Data using K-Nearest Neighbors}
\usage{
constructGraph(coordinates, k)
}
\arguments{
\item{coordinates}{Data frame whose first 2 columns are numeric coordinate data}

\item{k}{Integer, the parameter for the K-Nearest Neighbors algorithm}
}
\value{
An object of class 'graph' from the igraph package
}
\description{
Construct a Graph Object from Coordinate Data using K-Nearest Neighbors
}
\examples{
set.seed(1)
coords = data.frame(lon = rnorm(50), lat = rnorm(50))
g = constructGraph(coords, 4)
plot(g, layout = as.matrix(coords), edge.arrow.mode = 0)
}
