% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{edgeBetweenClust}
\alias{edgeBetweenClust}
\title{Get Betweenness of Edges in Clustered Graph}
\usage{
edgeBetweenClust(graph, membership)
}
\arguments{
\item{graph}{An object of class 'graph' from the igraph package.}

\item{membership}{A vector of integers of length N (see \code{\link{constructClusters}})}
}
\value{
A vector of boolean values of length \code{ecount(graph)} where each TRUE value represents an edge that connects distinct clusters
}
\description{
Get Betweenness of Edges in Clustered Graph
}
\examples{
set.seed(1)
coords = data.frame(lon = rnorm(50), lat = rnorm(50))
g = constructGraph(coords, 5)
clust_membership = constructClusters(g, 6, minclust = 5)$membership
betweenness = edgeBetweenClust(g, clust_membership)
plot(g, layout = as.matrix(coords), edge.color = as.numeric(betweenness) + 1, edge.arrow.mode = 0)
}
