% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{fit_BASTION}
\alias{fit_BASTION}
\title{Decomposition of time series data with Bayesian Adaptive Seasonality Trend Decomposition Incorporating Outlier and Noise (BASION)}
\usage{
fit_BASTION(
  y,
  Ks,
  X = NULL,
  Outlier = FALSE,
  cl = 0.95,
  sparse = FALSE,
  obsSV = "const",
  nchains = 2,
  nsave = 1000,
  nburn = 1000,
  nskip = 4,
  verbose = TRUE
)
}
\arguments{
\item{y}{numeric vector of the \code{T x 1} vector of time series observations}

\item{Ks}{list of values containing the seasonal periods}

\item{X}{matrix for additional covariate for regression (default is NULL)}

\item{Outlier}{logical flag (default is FALSE) to model outlier}

\item{cl}{scalar between (0,1) for confidence leve (default is 0.95)}

\item{sparse}{logical flag (default is FALSE) to induce additional shrinkage to the trend estimate}

\item{obsSV}{options for modeling the error variance. It must be one of the following:
\itemize{
\item const: Constant error variance for all time points.
\item SV: Stochastic Volatility model.
}}

\item{nchains}{integer scalar for the number of chains for MCMC sampling (default is 2)}

\item{nsave}{integer scalar (default = 1000); number of MCMC iterations to record}

\item{nburn}{integer scalar (default = 1000); number of MCMC iterations to discard (burn-in)}

\item{nskip}{integer scalar (default = 4); number of MCMC iterations to skip between saving iterations,
i.e., save every (nskip + 1)th draw}

\item{verbose}{logical; should R report extra information on progress? Defaults to FALSE}
}
\value{
\code{fit_BASTION} returns an object class list.
}
\description{
decompose time series data using BASTION. The time series should be a vector.
}
\section{\code{summary}}{

A list providing summarized posterior estimates:
\itemize{
\item \code{p_means}: A matrix containing the posterior means of the observed data (\code{y}) and each decomposed component:
\strong{Trend}, \strong{Seasonality}, and \strong{Outlier}, if applicable.
\item \code{Trend_sum}: A matrix containing the posterior mean and 95\% credible interval of the trend estimate.
\item \verb{Seasonal"k"_sum}: A matrix containing the posterior mean and 95\% credible interval of the seasonal estimate,
where \code{"k"} is determined by the input.
\item \code{Outlier_sum}: A matrix containing the posterior mean and 95\% credible interval of the outlier component.
}
}

\section{\code{samples}}{

A list containing MCMC samples of the relevant parameters:
\itemize{
\item \code{beta_combined}: Posterior samples of \strong{Trend + Seasonality}.
\item \code{beta}: Posterior samples of each component excluding the remainder.
\item \code{obs_sigma_t2}: Posterior samples of the variance of the observation equation.
\item \code{evol_sigma_t2}: Posterior samples of the evolution error term.
\item \code{remainder}: Posterior samples of the remainder term.
\item \code{yhat}: Posterior samples of the signal + error term.
}
}

