% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genMesh.R
\name{genMesh}
\alias{genMesh}
\title{Generate a random mesh inside an SFC boundary}
\usage{
genMesh(sfc_bnd, coords = NULL, n_ref = 100, n_bnd = NULL, ...)
}
\arguments{
\item{sfc_bnd}{An SFC object, the boundary inside which the mesh should be generated}

\item{coords}{Optional set of interior coordinate points for the mesh to span. By default, n_ref points are uniformly sampled using st_sample}

\item{n_ref}{Number of interior mesh nodes to randomly generate. Larger numbers result in a finer mesh}

\item{...}{}
}
\value{
An SF MULTIPOLYGON object
}
\description{
Generate a random mesh inside an SFC boundary
}
