% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polygon_constant.R
\name{get_polygon_constant}
\alias{get_polygon_constant}
\title{Get the piecewise constant value of each polygon}
\usage{
get_polygon_constant(membership, mus)
}
\arguments{
\item{membership}{A vector with length equal to the number of polygons, the output of igraph's components function}

\item{mus}{A vector with length equal to the number of distinct clusters, containing the numeric piecewise constant for each cluster}
}
\value{
A vector with length equal to the number of polygons, with each value being the constant of the cluster the polygon belongs to
}
\description{
Get the piecewise constant value of each polygon
}
