% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polygon_id.R
\name{get_polygon_id}
\alias{get_polygon_id}
\title{Get polygon ids of a data set}
\usage{
get_polygon_id(point_in_polygon)
}
\arguments{
\item{point_in_polygon}{The output of st_contains on a mesh and data set}
}
\value{
A data frame mapping each data point to the id of the polygon in which it is contained
}
\description{
Get polygon ids of a data set
}
