% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{graphBirth}
\alias{graphBirth}
\title{Perform a cluster birth operation (split an existing cluster)}
\usage{
graphBirth(graph, membership, clust = NULL)
}
\arguments{
\item{graph}{An object of class 'graph' from the igraph package}

\item{membership}{A vector of integers of length N with k unique integers (k < N) which map each vertex to a cluster}

\item{clust}{(Optional) Integer, the cluster to split. Must be between 1 and k. By default, a cluster will be uniformly randomly selected}
}
\value{
A list containing two elements:
'graph': The input graph with 1 fewer active edge
'membership': A vector of integers of length N with k + 1 unique integers which map each vertex to a cluster
}
\description{
Perform a cluster birth operation (split an existing cluster)
}
\examples{
set.seed(1234)
coords = data.frame(lon = rnorm(50), lat = rnorm(50))
g = constructGraph(coords, 5)
clust_out = constructClusters(g, 6, minclust = 5)
plot(clust_out$spanning_forest, layout = as.matrix(coords), vertex.color = clust_out$membership, edge.arrow.mode = 0)
g_7_clusters = graphBirth(clust_out$spanning_forest, clust_out$membership, 4)
plot(g_7_clusters$graph, layout = as.matrix(coords), vertex.color = g_7_clusters$membership, edge.arrow.mode = 0)
}
