% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{graphChange}
\alias{graphChange}
\title{Perform a cluster death operation followed by a cluster birth operation}
\usage{
graphChange(graph, membership, full_graph)
}
\arguments{
\item{graph}{An object of class 'graph' from the igraph package}

\item{membership}{A vector of integers of length N with k unique integers (1 < k <= N) which map each vertex to a cluster}

\item{full_graph}{An object of class 'graph' from the igraph package; 'graph' should be a subgraph of full_graph}
}
\value{
A list containing two elements:
\item{graph}{The input graph with the different set of active edges}
\item{membership}{A vector of integers of length N with k unique integers which map each vertex to a cluster, likely different than input membership}
}
\description{
Perform a cluster death operation followed by a cluster birth operation
}
\examples{
set.seed(1)
coords = data.frame(lon = rnorm(80), lat = rnorm(80))
g = constructGraph(coords, 5)
clust_out = constructClusters(g, 6, minclust = 8)
plot(clust_out$spanning_forest, layout = as.matrix(coords), vertex.color = clust_out$membership, edge.arrow.mode = 0)
g_different = graphChange(clust_out$spanning_forest, clust_out$membership, g)
plot(g_different$graph, , layout = as.matrix(coords), vertex.color = g_different$membership, edge.arrow.mode = 0)
}
