% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{graphDeath}
\alias{graphDeath}
\title{Perform a cluster death operation (merge an existing cluster)}
\usage{
graphDeath(graph, membership, full_graph)
}
\arguments{
\item{graph}{An object of class 'graph' from the igraph package}

\item{membership}{A vector of integers of length N with k unique integers (1 < k <= N) which map each vertex to a cluster}

\item{full_graph}{An object of class 'graph' from the igraph package; 'graph' should be a subgraph of full_graph}
}
\value{
A list containing two elements:
'graph': The input graph with 1 additional active edge
'membership': A vector of integers of length N with k - 1 unique integers which map each vertex to a cluster
}
\description{
Perform a cluster death operation (merge an existing cluster)
}
\examples{
set.seed(1)
coords = data.frame(lon = rnorm(50), lat = rnorm(50))
g = constructGraph(coords, 6)
clust_out = constructClusters(g, 8, minclust = 3)
plot(clust_out$spanning_forest, layout = as.matrix(coords), vertex.color = clust_out$membership, edge.arrow.mode = 0)
g_7_clusters = graphDeath(clust_out$spanning_forest, clust_out$membership, g)
plot(g_7_clusters$graph, layout = as.matrix(coords), vertex.color = g_7_clusters$membership, edge.arrow.mode = 0)
}
