% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{graphHyper}
\alias{graphHyper}
\title{Generate new spanning trees for each cluster}
\usage{
graphHyper(full_graph, membership)
}
\arguments{
\item{full_graph}{An object of class 'graph' from the igraph package}

\item{membership}{A vector of integers of length N with k unique integers (1 < k <= N) which map each vertex to a cluster}
}
\value{
A list containing two elements:
'graph': The input graph with each cluster having a new minimum spanning tree based on resampled edge weights
'membership': A vector of integers of length N with k unique integers which map each vertex to a cluster
}
\description{
Generate new spanning trees for each cluster
}
\examples{
set.seed(1)
coords = data.frame(lon = rnorm(100), lat = rnorm(100))
g = constructGraph(coords, 5)
clust_out = constructClusters(g, 8, minclust = 6)
plot(clust_out$spanning_forest, layout = as.matrix(coords), vertex.color = clust_out$membership, edge.arrow.mode = 0)
g_resample = graphHyper(g, clust_out$membership)
plot(g_resample$graph, layout = as.matrix(coords), vertex.color = g_resample$membership, edge.arrow.mode = 0)
}
