% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterGraph.R
\name{clusterGraph}
\alias{clusterGraph}
\title{Remove inter-cluster edges from a graph}
\usage{
clusterGraph(graph, membership)
}
\arguments{
\item{graph}{An object of class 'graph' from the \code{\link[igraph]{igraph}} package.}

\item{membership}{A vector of integers of length N (see \code{\link{constructClusters}})}
}
\value{
An object of class 'graph' from the \code{\link[igraph]{igraph}} package, the input graph with inter-cluster edges removed
}
\description{
\code{clusterGraph} takes in a graph and the membership vector mapping each vertex to
a certain cluster and returns the input graph with any edges that connect vertices belonging
to distinct clusters removed (see \code{\link{edgeBetweenClust}}).
}
\examples{
set.seed(1)
coords = data.frame(lon = rnorm(50), lat = rnorm(50))
g = constructGraph(coords, 4)
clust_membership = constructClusters(g, 5, minclust = 3)$membership
clust_g = clusterGraph(g, clust_membership)
plot(clust_g, layout = as.matrix(coords), vertex.color = clust_membership, edge.arrow.mode = 0)
}
\references{
Luo, Z.T. (*), Sang, H. and Mallick, B.K. (2021), BAST: Bayesian Additive Regression Spanning Trees
for Complex Constrained Domain

Luo, Z.T. (*), Sang, H. and Mallick, B.K. (2021), A Bayesian Contiguous Partitioning Method for
Learning Clustered Latent Variables, Journal of Machine Learning Research, 22, 1-52.
}
