% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgeBetweenClust.R
\name{edgeBetweenClust}
\alias{edgeBetweenClust}
\title{Get Betweenness of Edges in Clustered Graph}
\usage{
edgeBetweenClust(graph, membership)
}
\arguments{
\item{graph}{An object of class 'graph' from the \code{\link[igraph]{igraph}} package.}

\item{membership}{A vector of integers of length N (see \code{\link{constructClusters}})}
}
\value{
A vector of boolean values of length \code{ecount(graph)} where each TRUE value represents an edge that connects distinct clusters
}
\description{
\code{edgeBetweenClust} is a utility function which takes in a graph and the membership vector
mapping each vertex to a certain cluster and returns a vector of boolean values. A value of 'TRUE'
means that the corresponding edge in that edge list position connects vertices which belong to
distinct clusters. A value of 'FALSE' means that the edge connects vertices which belong to the same
cluster.
}
\examples{
set.seed(1)
coords = data.frame(lon = rnorm(50), lat = rnorm(50))
g = constructGraph(coords, 5)
clust_membership = constructClusters(g, 6, minclust = 5)$membership
betweenness = edgeBetweenClust(g, clust_membership)
plot(g,
     layout = as.matrix(coords),
     edge.color = as.numeric(betweenness) + 1,
     edge.arrow.mode = 0)
}
\references{
Luo, Z.T. (*), Sang, H. and Mallick, B.K. (2021), BAST: Bayesian Additive Regression Spanning Trees
for Complex Constrained Domain

Luo, Z.T. (*), Sang, H. and Mallick, B.K. (2021), A Bayesian Contiguous Partitioning Method for
Learning Clustered Latent Variables, Journal of Machine Learning Research, 22, 1-52.
}
